scripts/dbsync/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
